Fix Skifio DAC u16->i32

+++ source/mcu/main/src/drivers/skifio.h
@@ -12,7 +12,7 @@
 #define SKIFIO_DOUT_SIZE 4
 
 typedef int32_t SkifioAin;
-typedef int16_t SkifioAout;
+typedef int32_t SkifioAout;
 
 typedef struct SkifioInput {
     SkifioAin adcs[SKIFIO_ADC_CHANNEL_COUNT];
